+++ linux-user/xtensa/target_mman.h
@@ -22,7 +22,7 @@
 
 /* arch/xtensa/include/asm/elf.h */
 #define ELF_ET_DYN_BASE \
-    TARGET_PAGE_ALIGN((1u << TARGET_VIRT_ADDR_SPACE_BITS) / 3)
+    ((size_t) TARGET_PAGE_ALIGN((1u << TARGET_VIRT_ADDR_SPACE_BITS) / 3))
 
 #include "../generic/target_mman.h"
